Increase performance for a mix of ops.
According to valgrind, this patch reduces the number of instructions executed by
nearly a factor of 10.

The straightforward algorithm for BFS requires a queue to operate.  The queue
allocates a large amount of memory, and the majority of the "query order" calls
were bottlenecked on queue allocation and deallocation.

The fix is to add a "bfsnext" member to each vertex, and form a linked list
across vertices that acts as the queue.  I am spending 8b per vertex in order to
avoid any memory allocation when performing BFS.

This is inline with what we already do.  To avoid a linear sweep over all
vertices for BFS, we already spend 16b/vertex.